Rename: 마이페이지 탭 컴포넌트 폴더 이동
