+++ src/pages/ChatPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import * as Tabs from '@radix-ui/react-tabs';
 import styled from 'styled-components';
 import JoinList from '../components/feature/chat/tabs/JoinList';
-import Settings from './mypage/tabs/Settings';
+import Settings from '../components/feature/mypage/tabs/Settings';
 // import ChatPageSideBar from '../components/feature/chat/ChatPageSideBar';
 // import ChatPageMain from '../components/feature/chat/ChatPageMain';
